修正一些warning

+++ TestCase008/Viewer.h
@@ -46,10 +46,10 @@ private:
 	std::shared_ptr<CGLFrameBuffer> m_pFrameBuffer = nullptr;
 
 	bool m_IsFirstMouse = true;
-	int m_DrawElementsCnt = 0;
+	unsigned int m_DrawElementsCnt = 0;
 	int m_WindowWidth = 0, m_WindowHeight = 0;
 	int m_FBO = -1;
-	float m_LastMouseX = 0.0f, m_LastMouseY = 0.0f;
+	double m_LastMouseX = 0.0f, m_LastMouseY = 0.0f;
 	double m_LastGLFWTime = 0.0f, m_DeltaTime = 0.0f;
 	float m_RotateSpeed = 1.0f;
 	std::deque<float> m_FrameDeque;
